+++ .eslintrc.js
@@ -1,0 +1,10 @@
+module.exports = {
+    extends: [
+        'socifi',
+    ],
+    rules: {
+        'no-plusplus': ['error', { allowForLoopAfterthoughts: true }],
+        'class-methods-use-this': 0,
+    },
+    root: true,
+};